Update README.md
Included `build` and `docs` directories to example structure. This project https://github.com/php-pds/skeleton shows `docs` as a root level directory, which could make sense to have beyond `README.md` and `CHANGELOG.md` files. The `build` directory is also within the `.gitignore` file of this project, so there's already an acknowledgement that it may be used.